pyproject.toml
[tool.poetry]
name = "store"
version = "0.1.0"
description = ""
authors = ["Zakhar Fesik <fesikzakhar@khmnu.edu.ua>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
pillow = "^10.0.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
psycopg2 = "^2.9.7"
cryptography = "41.0.0"
django-allauth = "^0.55.0"
django-debug-toolbar = "^4.2.0"
redis = "^5.0.0"
celery = {extras = ["redis"], version = "^5.3.1"}
stripe = "^6.3.0"
ruff = "^0.0.287"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"

[tool.ruff]
exclude = ['migrations']
line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
